+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Maybe: Merge watched posts and watched stories to one setting
 
+## [2.7.5] - 2018-06-26
+
+### Bug Fixes
+
+- No recent items were added to the 3-dots menu.
+
 ## [2.7.3] - 2018-06-24
 
 ### Added